chore(vscode): organize files under treeview
