+++ packages/vscode-extension/src/treeview/treeViewManager.ts
@@ -5,8 +5,8 @@
 import * as vscode from "vscode";
 import * as exp from "../exp";
 import { TreatmentVariables } from "../exp/treatmentVariables";
-import { CommandsWebviewProvider } from "./commandsWebviewProvider";
-import { TreeContainerType } from "./treeContainerType";
+import { CommandsWebviewProvider } from "./webViewProvider/commandsWebviewProvider";
+import { TreeContainerType } from "./webViewProvider/treeContainerType";
 import * as StringResources from "../resources/Strings.json";
 import { CommandsTreeViewProvider, TreeViewCommand } from "./commandsTreeViewProvider";
 import { TreeCategory } from "@microsoft/teamsfx-api";
