+++ packages/vscode-extension/src/treeview/webViewProvider/deploymentView.tsx
@@ -4,7 +4,7 @@ import publish_dark from "../../media/dark/publish.svg";
 import publish_light from "../../media/light/publish.svg";
 import developerPortal_dark from "../../media/dark/developerPortal.svg";
 import developerPortal_light from "../../media/light/developerPortal.svg";
-import * as StringResources from "../resources/Strings.json";
+import * as StringResources from "../../resources/Strings.json";
 import { getCurrentTheme, Theme } from "./theme";
 import { TreeItem } from "./treeItem";
