+++ packages/vscode-extension/src/treeview/webViewProvider/tree.scss
